fix after PR comments
