+++ helpers/downloadNotes.js
@@ -1,35 +1,19 @@
-import { useTranslation } from 'next-i18next'
+import axios from 'axios'
+
 const downloadNotes = async (check) => {
-  const { t } = useTranslation
-  try {
-    const responseMaterials = await fetch(`/api/checks/${check.id}/notes`)
-    const notes = await responseMaterials.json()
-
-    const notesDownload = ['chapter\tverse\tnote']
-    if (!notes.error) {
-      notes.forEach((line) => {
-        const { chapter, verse, note } = line
-        const row = `${chapter}\t${verse}\t${note}`
-        notesDownload.push(row)
-      })
-
-      const tsvContent = notesDownload.join('\n')
-
-      const blob = new Blob([tsvContent])
-
-      const link = document.createElement('a')
-      link.href = window.URL.createObjectURL(blob)
-      link.download = `${check.name}.tsv`
-
-      document.body.appendChild(link)
-      link.click()
-      document.body.removeChild(link)
+  return await axios.get(`/api/checks/${check.id}/notes`).then((responseMaterials) => {
+    const notes = responseMaterials.data
+    if (notes.error) {
+      throw error
     }
-    return notesDownload
-  } catch (error) {
-    console.error(t('errorDownloadNotes'), error)
-    throw error
-  }
+    const notesDownload = ['chapter\tverse\tnote']
+    notes.forEach((line) => {
+      const { chapter, verse, note } = line
+      const row = `${chapter}\t${verse}\t${note}`
+      notesDownload.push(row)
+    })
+    return notesDownload.join('\n')
+  })
 }
 
 export default downloadNotes
