+++ public/locales/en/common.json
@@ -1,53 +1,56 @@
 {
-    "accept": "Accept",
-    "activity": "Activity (number of checks)",
-    "activeChecks": "Active checks",
-    "back": "Back",
-    "bookChecks": "Book checks",
-    "bookTitle": "Book title",
-    "checkEndDate": "Check end date",
-    "checkDetails": "Check details",
-    "copied": "Copied!",
-    "create": "Create",
-    "createProject": "Create a project",
-    "createBook": "Create a book",
-    "dateCreation": "Date of creation",
-    "dateLastCheck": "Date of last check",
-    "download": "Download",
-    "uploadMaterial": "Upload material",
-    "downloadNotes": "Download notes",
-    "editProject": "Edit project",
-    "email": "Email",
-    "enterCorrectLink": "Enter the correct resource link",
-    "errorOccurred": "An error occurred",
-    "errorDownloadNotes": "Error downloading notes:",
-    "expirationDate": "Expiration date:",
-    "goToProjects": "Go to projects",
-    "linkResource": "Link to resource",
-    "loading": "Loading",
-    "logout": "Logout",
-    "name": "Name:",
-    "nextChapter": "Next сhapter",
-    "note": "Note",
-    "noteSaved": "Note saved",
-    "number": "Number",
-    "numberChecks": "Number of checks",
-    "password": "Password",
-    "personalArea": "Personal Area",
-    "previousChapter": "Previous chapter",
-    "projectBooks": "Project books",
-    "projects": "Projects",
-    "provideLink": "Provide a link to the material:",
-    "save": "Save",
-    "signIn": "Sign in",
-    "signAccount": "Sign in to your account",
-    "signOut": "Sign Out",
-    "startNewCheck": "Start a new check",
-    "titleInTable": "Title",
-    "titleProject": "Title project",
-    "nameEmpty": "Title cannot be empty",
-    "welcomeTo": "Welcome to Community Check",
-    "youLoggedIn": "You are already logged in.",
-    "loginToAccess": "Please login to access full functionality.",
-    "createCheck": "Create a Check"
+  "accept": "Accept",
+  "activeChecks": "Active checks",
+  "activity": "Activity (number of checks)",
+  "back": "Back",
+  "bookChecks": "Book checks",
+  "bookTitle": "Book title",
+  "checkDetails": "Check details",
+  "checkEndDate": "Check end date",
+  "copied": "Copied!",
+  "create": "Create",
+  "createBook": "Create a book",
+  "createCheck": "Create a Check",
+  "createProject": "Create a project",
+  "dateCreation": "Date of creation",
+  "dateLastCheck": "Date of last check",
+  "download": "Download",
+  "downloadNotes": "Download notes",
+  "editProject": "Edit project",
+  "email": "Email",
+  "enterCorrectLink": "Enter the correct resource link",
+  "errorDownloadNotes": "Error downloading notes:",
+  "errorOccurred": "An error occurred",
+  "expirationDate": "Expiration date:",
+  "goToProjects": "Go to projects",
+  "linkResource": "Link to resource",
+  "loading": "Loading",
+  "loginToAccess": "Please login to access full functionality.",
+  "logout": "Logout",
+  "messageAboutNewCheck": "Your check was created with a default name. Refresh the page",
+  "name": "Name:",
+  "nameCheck": "Name Check",
+  "nameEmpty": "Title cannot be empty",
+  "nextChapter": "Next сhapter",
+  "note": "Note",
+  "noteSaved": "Note saved",
+  "number": "Number",
+  "numberChecks": "Number of checks",
+  "password": "Password",
+  "personalArea": "Personal Area",
+  "previousChapter": "Previous chapter",
+  "projectBooks": "Project books",
+  "projects": "Projects",
+  "provideLink": "Provide a link to the material:",
+  "register": "Register",
+  "save": "Save",
+  "signAccount": "Sign in to your account",
+  "signIn": "Sign in",
+  "signOut": "Sign Out",
+  "startNewCheck": "Start a new check",
+  "titleInTable": "Title",
+  "titleProject": "Title project",
+  "uploadMaterial": "Upload material",
+  "welcomeTo": "Welcome to Community Check",
+  "youLoggedIn": "You are already logged in."
 }
