+++ public/locales/ru/common.json
@@ -1,53 +1,56 @@
 {
-    "accept": "Принять",
-    "activity": "Активность (количество проверок)",
-    "activeChecks": "Активные проверки",
-    "back": "Назад",
-    "bookChecks": "Проверки книги",
-    "bookTitle": "Заголовок книги",
-    "checkEndDate": "Дата завершения проверки",
-    "checkDetails": "Детали проверки",
-    "copied": "Скопировано!",
-    "create": "Создать",
-    "createProject": "Создать проект",
-    "createBook": "Создать книгу",
-    "dateCreation": "Дата создания",
-    "dateLastCheck": "Дата последней проверки",
-    "download": "Скачать",
-    "uploadMaterial": "Загрузить материал",
-    "downloadNotes": "Скачать заметки",
-    "editProject": "Редактировать проект",
-    "email": "Электронная почта",
-    "enterCorrectLink": "Введите правильную ссылку на ресурс",
-    "errorOccurred": "Произошла ошибка",
-    "errorDownloadNotes": "Ошибка при скачивании заметок:",
-    "expirationDate": "Дата истечения срока действия:",
-    "goToProjects": "Перейти к проектам",
-    "linkResource": "Ссылка на ресурс",
-    "loading": "Загрузка",
-    "logout": "Выйти",
-    "name": "Название:",
-    "number": "Номер",
-    "numberChecks": "Количество проверок",
-    "nextChapter": "Следующая глава",
-    "note": "Заметка",
-    "noteSaved": "Заметка сохранена",
-    "password": "Пароль",
-    "personalArea": "Личный кабинет",
-    "projectBooks": "Книги проекта",
-    "projects": "Проекты",
-    "provideLink": "Укажите ссылку на материал:",
-    "previousChapter": "Предыдущая глава",
-    "save": "Сохранить",
-    "signIn": "Войти",
-    "signAccount": "Войти в аккаунт",
-    "signOut": "Выйти",
-    "startNewCheck": "Начать новую проверку",
-    "titleInTable": "Заголовок",
-    "titleProject": "Название проекта",
-    "nameEmpty": "Название не может быть пустым",
-    "welcomeTo": "Добро пожаловать в Community Check",
-    "youLoggedIn": "Вы уже вошли в систему.",
-    "loginToAccess": "Пожалуйста, войдите, чтобы получить доступ к полному функционалу.",
-    "createCheck": "Создать проверку"
+	"accept": "Принять",
+	"activeChecks": "Активные проверки",
+	"activity": "Активность (количество проверок)",
+	"back": "Назад",
+	"bookChecks": "Проверки книги",
+	"bookTitle": "Заголовок книги",
+	"checkDetails": "Детали проверки",
+	"checkEndDate": "Дата завершения проверки",
+	"copied": "Скопировано!",
+	"create": "Создать",
+	"createBook": "Создать книгу",
+	"createCheck": "Создать проверку",
+	"createProject": "Создать проект",
+	"dateCreation": "Дата создания",
+	"dateLastCheck": "Дата последней проверки",
+	"download": "Скачать",
+	"downloadNotes": "Скачать заметки",
+	"editProject": "Редактировать проект",
+	"email": "Электронная почта",
+	"enterCorrectLink": "Введите правильную ссылку на ресурс",
+	"errorDownloadNotes": "Ошибка при скачивании заметок:",
+	"errorOccurred": "Произошла ошибка",
+	"expirationDate": "Дата истечения срока действия:",
+	"goToProjects": "Перейти к проектам",
+	"linkResource": "Ссылка на ресурс",
+	"loading": "Загрузка",
+	"loginToAccess": "Пожалуйста, войдите, чтобы получить доступ к полному функционалу.",
+	"logout": "Выйти",
+	"messageAboutNewCheck": "Ваша проверка создана с именем по умолчанию. Обновите страницу.",
+	"name": "Название:",
+	"nameCheck": "Название Проверки",
+	"nameEmpty": "Название не может быть пустым",
+	"nextChapter": "Следующая глава",
+	"note": "Заметка",
+	"noteSaved": "Заметка сохранена",
+	"number": "Номер",
+	"numberChecks": "Количество проверок",
+	"password": "Пароль",
+	"personalArea": "Личный кабинет",
+	"previousChapter": "Предыдущая глава",
+	"projectBooks": "Книги проекта",
+	"projects": "Проекты",
+	"provideLink": "Укажите ссылку на материал:",
+	"register": "Регистрация",
+	"save": "Сохранить",
+	"signAccount": "Войти в аккаунт",
+	"signIn": "Войти",
+	"signOut": "Выйти",
+	"startNewCheck": "Начать новую проверку",
+	"titleInTable": "Заголовок",
+	"titleProject": "Название проекта",
+	"uploadMaterial": "Загрузить материал",
+	"welcomeTo": "Добро пожаловать в Community Check",
+	"youLoggedIn": "Вы уже вошли в систему."
 }
